+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ENHANCEMENTS:
 
 * Add `--output` flag to `replace` command to output as line diffs for each replacement in addition to the default inline format. ([#88](https://github.com/fishi0x01/vsh/pull/88))
+* Add `add` command for single key insertion ([#87](https://github.com/fishi0x01/vsh/pull/87))
 
 BUG FIXES: